clean up some code and modules to improve debug speed
