+++ web/components/shared/upgradeProModal.tsx
@@ -1,16 +1,18 @@
-import { SparklesIcon } from "@heroicons/react/24/outline";
+// import { SparklesIcon } from "@heroicons/react/24/outline";
 import { useUser } from "@supabase/auth-helpers-react";
 import { useState } from "react";
 import { useOrg } from "../layout/organizationContext";
 import ThemedModal from "./themed/themedModal";
 import getStripe from "../../utlis/getStripe";
-import {
-  ORGANIZATION_COLORS,
-  ORGANIZATION_ICONS,
-} from "../templates/organization/createOrgForm";
+
 import { clsx } from "./clsx";
 import { useGetRequestCountClickhouse } from "../../services/hooks/requests";
 import { endOfMonth, formatISO, startOfMonth } from "date-fns";
+import {
+  ORGANIZATION_COLORS,
+  ORGANIZATION_ICONS,
+} from "../templates/organization/orgConstants";
+import { SparklesIcon } from "lucide-react";
 
 interface UpgradeProModalProps {
   open: boolean;
