+++ web/components/templates/organization/createOrgForm.tsx
@@ -1,117 +1,16 @@
 import { RadioGroup } from "@headlessui/react";
-import {
-  BuildingOfficeIcon,
-  CakeIcon,
-  CloudIcon,
-  CommandLineIcon,
-  RocketLaunchIcon,
-  CpuChipIcon,
-  ScaleIcon,
-  Square3Stack3DIcon,
-} from "@heroicons/react/24/outline";
-import { useSupabaseClient, useUser } from "@supabase/auth-helpers-react";
+
+import { useUser } from "@supabase/auth-helpers-react";
 import { useState } from "react";
+import { getJawnClient } from "../../../lib/clients/jawn";
 import { DEMO_EMAIL } from "../../../lib/constants";
-import { Database } from "../../../supabase/database.types";
-import { clsx } from "../../shared/clsx";
 import { useOrg } from "../../layout/organizationContext";
+import { clsx } from "../../shared/clsx";
 import useNotification from "../../shared/notification/useNotification";
 import ProviderKeyList from "../enterprise/portal/id/providerKeyList";
 import CreateProviderKeyModal from "../vault/createProviderKeyModal";
 import { useVaultPage } from "../vault/useVaultPage";
-import { getJawnClient } from "../../../lib/clients/jawn";
-
-export const ORGANIZATION_COLORS = [
-  {
-    name: "gray",
-    bgColor: "bg-gray-200",
-    textColor: "text-gray-800",
-    selectedColor: "ring-gray-500",
-  },
-  {
-    name: "red",
-    bgColor: "bg-red-300",
-    textColor: "text-red-800",
-    selectedColor: "ring-red-500",
-  },
-  {
-    name: "yellow",
-    bgColor: "bg-yellow-200",
-    textColor: "text-yellow-800",
-    selectedColor: "ring-yellow-500",
-  },
-  {
-    name: "green",
-    bgColor: "bg-green-300",
-    textColor: "text-green-800",
-    selectedColor: "ring-green-500",
-  },
-  {
-    name: "blue",
-    bgColor: "bg-blue-300",
-    textColor: "text-blue-800",
-    selectedColor: "ring-blue-500",
-  },
-  {
-    name: "purple",
-    bgColor: "bg-purple-300",
-    textColor: "text-purple-800",
-    selectedColor: "ring-purple-500",
-  },
-];
-
-type OrgIconType = {
-  name:
-    | "building"
-    | "cake"
-    | "cloud"
-    | "rocket"
-    | "code"
-    | "chip"
-    | "scale"
-    | "stack";
-  icon: React.ForwardRefExoticComponent<
-    React.SVGProps<SVGSVGElement> & {
-      title?: string | undefined;
-      titleId?: string | undefined;
-    }
-  >;
-};
-
-export const ORGANIZATION_ICONS: OrgIconType[] = [
-  {
-    name: "building",
-    icon: BuildingOfficeIcon,
-  },
-  {
-    name: "cake",
-    icon: CakeIcon,
-  },
-  {
-    name: "cloud",
-    icon: CloudIcon,
-  },
-  {
-    name: "rocket",
-    icon: RocketLaunchIcon,
-  },
-  {
-    name: "code",
-    icon: CommandLineIcon,
-  },
-  {
-    name: "chip",
-    icon: CpuChipIcon,
-  },
-  {
-    name: "scale",
-    icon: ScaleIcon,
-  },
-  {
-    name: "stack",
-    icon: Square3Stack3DIcon,
-  },
-];
+import { ORGANIZATION_COLORS, ORGANIZATION_ICONS } from "./orgConstants";
 
 interface CreateOrgFormProps {
   variant?: "organization" | "reseller";
@@ -170,13 +69,12 @@ const CreateOrgForm = (props: CreateOrgFormProps) => {
   const user = useUser();
   const orgContext = useOrg();
   const { setNotification } = useNotification();
-  const supabaseClient = useSupabaseClient<Database>();
   const [providerKey, setProviderKey] = useState(
     initialValues?.providerKey || ""
   );
 
   const { providerKeys, refetchProviderKeys } = useVaultPage();
-  const [deleteOpen, setDeleteOpen] = useState(false);
+
   const [isProviderOpen, setIsProviderOpen] = useState(false);
 
   return (
